+++ NEWS.md
@@ -1,3 +1,8 @@
-# facetious 0.1.0
+# facetious 0.1.1   2021-07-21
+
+* `facet_grid_blank()` now allows for custom grobs in empty facet locations
+
+
+# facetious 0.1.0   2021-07-20
 
 * Initial release